SpellBook Token uses CharacterDisplay
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@16166 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/cdom/facet/SpellBookFacet.java
@@ -197,7 +197,7 @@ public class SpellBookFacet extends AbstractStorageFacet implements
 		return (componentMap != null) && componentMap.containsKey(name);
 	}
 
-	public double getCount(CharID id)
+	public int getCount(CharID id)
 	{
 		Map<String, SpellBook> componentMap = getCachedMap(id);
 		return (componentMap == null) ? 0 : componentMap.size();
